refactor: move tests into their own corresponding files
